// examples/next-js-sdk-gen-2-experimental-app-directory/src/sdk/blocks/symbol/symbol.types.ts
import type { BuilderContent } from '../../types/builder-content';
import type { BuilderComponentsProp } from '../../types/builder-props';
interface SymbolInfo {
  model?: string;
  entry?: string;
  data?: any;
  content?: BuilderContent;
  inline?: boolean;
  dynamic?: boolean;
}
export interface SymbolProps extends BuilderComponentsProp {
  symbol?: SymbolInfo;
  dataOnly?: boolean;
  dynamic?: boolean;
  attributes?: any;
  inheritState?: boolean;
}